increase the maximum size of query results to 2^64
Currently, we cannot select more than 2^32 rows from a table because we are limited by types of
variables containing the numbers of rows. This patch changes these types and sets new limits.

The new limits take effect while selecting all rows from a table - custom limits of rows in a result
stay the same (2^32-1).

In classes which are being serialized and used in messaging, in order to be able to process queries
originating from older nodes, the top 32 bits of new integers are optional and stay at the end
of the class - if they're not present, we assume they equal 0.

Fixes #5101.

+++ test/lib/random_schema.cc
@@ -1096,7 +1096,7 @@ future<std::vector<mutation>> generate_random_mutations(
 
             auto ckeys = random_schema.make_ckeys(clustering_row_count_dist(engine));
             const auto clustering_row_count = ckeys.size();
-            for (uint32_t ck = 0; ck < clustering_row_count; ++ck) {
+            for (uint64_t ck = 0; ck < clustering_row_count; ++ck) {
                 random_schema.add_row(engine, mut, ckeys[ck], ts_gen, exp_gen);
             }
